Reuse `EventEnvelope` in `Events.getTenantId(...)`
